requirements.txt
streamlit
fastapi
uvicorn
requests
langchain
langchain-openai
langchain-groq
langchain-community
tavily-python
python-dotenv

# Added for PDF reading
pypdf
unstructured
chardet